Fin de la pregunta 3

+++ DSWII_T3_YanezOmar/src/main/java/com/examen/DSWII_T3_YanezOmar/model/bd/Usuario.java
@@ -11,7 +11,7 @@ import java.util.Set;
 public class Usuario {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer idusuario;
+    private Long idusuario;
     @Column(name = "nomusuario")
     private String nomusuario;
     @Column(name = "email")
